requirements.txt
asgiref==3.2.10
autopep8==1.5.4
dj-database-url==0.5.0
Django==2.1.15
django-adminplus==0.5
django-crud-generator==0.3.10
django-filter==2.3.0
django-modeladmin-reorder==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
Owlready2==0.24
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
python-gettext==4.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
whitenoise==5.2.0
zipp==3.1.0
